Remove pbjs.requestAllBids() and refactor out adapter configs from prebid.js into adaptermanager.js
