+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "Prebid.js",
-  "version": "0.2.1",
+  "version": "0.3.1",
   "description": "Client Side Auction Library",
   "main": "prebid.js",
   "scripts": {},
