+++ src/constants.json
@@ -19,5 +19,10 @@
 			"ALL_BIDS_BACK" : "allRequestedBidsBack",
 			"AD_UNIT_BIDS_BACK" : "adUnitBidsBack"
 		}
-	}
+	},
+	"objectType_function" : "function",
+	"objectType_undefined" : "undefined",
+	"objectType_object" : "object",
+	"objectType_string" : "string",
+	"objectType_number" : "number"
 }
